Assignment 3 final changes

+++ Assignment3/test/input2.c
@@ -2,7 +2,7 @@ typedef float A;
 
 int main(){
     A b=3;
-    double c=b;
-    long d=4;
+    float c=b;
+    int d=4;
     return 0;
 }
